hw10 - исправления по замечаниям
