+++ hw10/backend/src/main/java/ru/otus/hw/rest/controller/AuthorRestController.java
@@ -3,7 +3,6 @@ package ru.otus.hw.rest.controller;
 
 import java.util.List;
 import lombok.AllArgsConstructor;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 import ru.otus.hw.dto.AuthorDTO;
@@ -15,12 +14,6 @@ public class AuthorRestController {
 
     private final AuthorService authorService;
 
-/*    @GetMapping("authors")
-    public ResponseEntity<List<AuthorDTO>> getAuthors() {
-        var authors = authorService.findAll();
-        return ResponseEntity.ok(authors);
-    }
-    */
 
     @GetMapping("/api/v1/authors")
     public List<AuthorDTO> getAllAuthors() {
